separate tests into 2 distinct files
