+++ tests/input_converter/test_converter.py
@@ -10,7 +10,7 @@
 #
 # This file is part of the Antares project.
 
-from typing import Callable, Literal
+from typing import Callable
 
 import pandas as pd
 import pytest
@@ -784,129 +784,3 @@ class TestConverter:
                 if thermal.area_id == "fr":
                     tdp = ThermalDataPreprocessing(thermal, study_path)
                     return tdp
-
-    def _setup_test(self, local_study_w_thermal, filename):
-        """
-        Initializes test parameters and returns the instance and expected file path.
-        """
-        areas, converter = self._init_area_reading(local_study_w_thermal)
-        study_path = converter.study_path
-        instance = self._generate_tdp_instance_parameter(areas, study_path)
-        expected_path = (
-            study_path / "input" / "thermal" / "series" / "fr" / "gaz" / filename
-        )
-        return instance, expected_path
-
-    def _validate_component(
-        self, instance, process_method, expected_path, expected_values
-    ):
-        """
-        Executes the given processing method, validates the component, and compares the output dataframe.
-        """
-        component = getattr(instance, process_method)()
-        expected_component = InputComponentParameter(
-            id=process_method.split("process_")[1],
-            time_dependent=True,
-            scenario_dependent=True,
-            value=str(expected_path),
-        )
-        current_df = pd.read_csv(expected_path.with_suffix(".txt"), header=None)
-        expected_df = pd.DataFrame(expected_values)
-        assert current_df.equals(expected_df)
-        assert component == expected_component
-
-    def _test_p_min_cluster(self, local_study_w_thermal):
-        """Tests the p_min_cluster parameter processing."""
-        instance, expected_path = self._setup_test(
-            local_study_w_thermal, "p_min_cluster.txt"
-        )
-        expected_values = [
-            [6.0],
-            [10.0],
-            [2.0],
-        ]  # min(min_gen_modulation * unit_count * nominal_capacity, p_max_cluster)
-        self._validate_component(
-            instance, "process_p_min_cluster", expected_path, expected_values
-        )
-
-    def test_nb_units_min(self, local_study_w_thermal: Study):
-        """Tests the nb_units_min parameter processing."""
-        instance, expected_path = self._setup_test(
-            local_study_w_thermal, "nb_units_min"
-        )
-        instance.process_p_min_cluster()
-        expected_values = [[2.0], [5.0], [1.0]]  # ceil(p_min_cluster / p_max_unit)
-        self._validate_component(
-            instance, "process_nb_units_min", expected_path, expected_values
-        )
-
-    def test_nb_units_max(self, local_study_w_thermal: Study):
-        """Tests the nb_units_max parameter processing."""
-        instance, expected_path = self._setup_test(
-            local_study_w_thermal, "nb_units_max"
-        )
-        instance.process_p_min_cluster()
-        expected_values = [[4.0], [5.0], [1.0]]  # ceil(p_max_cluster / p_max_unit)
-        self._validate_component(
-            instance, "process_nb_units_max", expected_path, expected_values
-        )
-
-    @pytest.mark.parametrize("direction", ["forward", "backward"])
-    def test_nb_units_max_variation(
-        self,
-        local_study_w_thermal: Study,
-        create_csv_from_constant_value: Callable[..., None],
-        direction: Literal["forward"] | Literal["backward"],
-    ):
-        """
-        Tests nb_units_max_variation_forward and nb_units_max_variation_backward processing.
-        """
-        instance, expected_path = self._setup_test(
-            local_study_w_thermal, f"nb_units_max_variation_{direction}"
-        )
-        modulation_timeseries = (
-            instance.study_path / "input" / "thermal" / "prepro" / "fr" / "gaz"
-        )
-        series_path = (
-            instance.study_path / "input" / "thermal" / "series" / "fr" / "gaz"
-        )
-        create_csv_from_constant_value(modulation_timeseries, "modulation", 840, 4)
-        create_csv_from_constant_value(series_path, "series", 840)
-        instance.process_nb_units_max()
-        nb_units_max_output = pd.read_csv(
-            instance.series_path / "nb_units_max.txt", header=None
-        )
-
-        variation_component = getattr(
-            instance, f"process_nb_units_max_variation_{direction}"
-        )()
-        current_df = pd.read_csv(variation_component.value + ".txt", header=None)
-
-        assert current_df[0][0] == max(
-            0, nb_units_max_output[0][167] - nb_units_max_output[0][0]
-        )
-        assert current_df[0][3] == max(
-            0, nb_units_max_output[0][2] - nb_units_max_output[0][3]
-        )
-        assert current_df[0][168] == max(
-            0, nb_units_max_output[0][335] - nb_units_max_output[0][168]
-        )
-        assert variation_component.value == str(expected_path)
-
-    def test_nb_units_max_variation_forward(
-        self,
-        local_study_w_thermal: Study,
-        create_csv_from_constant_value: Callable[..., None],
-    ):
-        self.test_nb_units_max_variation(
-            local_study_w_thermal, create_csv_from_constant_value, direction="forward"
-        )
-
-    def test_nb_units_max_variation_backward(
-        self,
-        local_study_w_thermal: Study,
-        create_csv_from_constant_value: Callable[..., None],
-    ):
-        self.test_nb_units_max_variation(
-            local_study_w_thermal, create_csv_from_constant_value, direction="backward"
-        )
